+ generalize types from radio button, filter, and puzzle input components + globalize DIFFICULTY type
