+++ src/types/filter.ts
@@ -1,0 +1,5 @@
+import { Dispatch, SetStateAction } from 'react';
+
+export type FilterProps = {
+  setFilterFields?: Dispatch<SetStateAction<object>>;
+};
