+++ src/types/puzzle.ts
@@ -1,0 +1,6 @@
+export type PuzzleInputProps = {
+  type: 'TEXT' | 'MCQ';
+  placeholder?: string;
+  options?: string[];
+  setAnswer: (value) => void;
+};
