+++ src/types/radioButton.ts
@@ -1,0 +1,9 @@
+import { DIFFICULTY } from './global';
+
+export type RadioButtonProps = {
+  id: string;
+  name: string;
+  difficulty: DIFFICULTY;
+  checked: boolean;
+  onClick?: () => void;
+};
